requirements.txt
flask==2.0.1
pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
opencv-python==4.8.1.78
numpy==1.24.4
